+++ aie-st-app/main.py
@@ -1,10 +1,11 @@
 import datetime
-from pathlib import Path
+import jwt
 import os
+from pathlib import Path
 import streamlit as st
+from streamlit.web.server.websocket_headers import _get_websocket_headers
 import sys
 import yaml
-from streamlit.web.server.websocket_headers import _get_websocket_headers
 
 
 class StApp:
@@ -14,12 +15,24 @@ class StApp:
         self.name = name_yaml
         self.path = "".join([str(self.here), "/", self.name])
         self.port = server_port
+        self.headers = _get_websocket_headers()
+        self.aad_user_jwt =  self.headers.get("X-Amzn-Oidc-Data")
+        
+    def decode_jwt(self, jwt_in: str) -> dict[str, str]:
+        """
+        Decode jwt from header key X-Amzn-Oidc-Data
+        """
+        decoded = jwt.decode(jwt_in, options={"verify_signature": False})
+        return decoded
+
 
     def main(self) -> int:
-        """Simple streamlit app that parse an input yaml and displays the content.
-           And also displays current user.
         """
-        
+        Simple streamlit app that parse an input yaml and displays the content.
+        And also displays current user.
+        """
+        user = self.decode_jwt(jwt_in=self.aad_user_jwt).get("email") 
+        st.title(f"Welcome {user}!")
         st.title('Hello World!')
         st.title("Current date and time: ")
         now = datetime.datetime.now()
@@ -38,20 +51,15 @@ class StApp:
             except yaml.YAMLError as exec:
                 print(exec)
                 raise
-        
-        st.title("======================================")
-        headers = _get_websocket_headers()
-        st.write("headers:")
-        st.write(headers)
-        if "X-Ms-Client-Principal-Name" in headers:
-            user_email = headers["X-Ms-Client-Principal-Name"]
-            st.write("user_email:")
-            st.write(user_email)
+        st.write("Web socket header data:")
+        st.write(self.headers)
 
         return 0
     
-    def main_cli(self) -> int:
-        """This function can be used by setup.py to create a shell command."""
+    def main_cli(self) -> None:
+        """
+        This function can be used by setup.py to create a shell command via entrypoint.
+        """
         os.system(f"streamlit run {__file__} --server.port {self.port}")
 
 if __name__ == "__main__":