+++ delivery-sheets/source/changelog.rst
@@ -1,10 +1,13 @@
 Change Log
 ==========
 
-- 2022-10-26
-    initial setup
+- 2022-11-30
+    implement personalization, by getting user data from web socket header
 
 - 2022-11-15
     implementation delivery sheets
+
+- 2022-10-26
+    initial setup